+++ shaders/inference/generalTranspose.wgsl
@@ -1,5 +1,5 @@
-@group(0) @binding(0) var<storage, read> input : array<f32>;
-        @group(0) @binding(1) var<storage, read_write> output : array<f32>;
+@group(0) @binding(0) var<storage, read> input : array<u32>;
+        @group(0) @binding(1) var<storage, read_write> output : array<u32>;
         @group(0) @binding(2) var<storage, read> md : array<u32>;
         @compute @workgroup_size(64)
         fn main(@builtin(global_invocation_id) gid : vec3<u32>) {
